Neural networks/requirements.txt
tensorflow
numpy
PyYAML
pandas
matplotlib
ucimlrepo
